+++ src/torchmetrics/classification/jaccard.py
@@ -172,7 +172,7 @@ class MulticlassJaccardIndex(MulticlassConfusionMatrix):
         self.average = average
 
     def compute(self) -> Tensor:
-        return _jaccard_index_reduce(self.confmat, average=self.average)
+        return _jaccard_index_reduce(self.confmat, average=self.average, ignore_index=self.ignore_index)
 
 
 class MultilabelJaccardIndex(MultilabelConfusionMatrix):